+++ docs/CHANGELOG.md
@@ -6,3 +6,4 @@
 - Do not ask for SRID when geometry is empty
 - Add the name of the table in the FGB header
 - Fix null way id on OSM file
+- Add ST_SnapToSelf function